feat(profile): add interaction states to account widget
