+++ src/pages/Profile/Profile.tsx
@@ -1,3 +1,8 @@
+import { useState } from 'react';
+
+// Components
+import { ConnectWalletButton } from 'containers';
+import { Tooltip } from 'components';
 import {
 	Link,
 	Redirect,
@@ -6,52 +11,34 @@ import {
 	useHistory,
 	useRouteMatch,
 } from 'react-router-dom';
-import { useWeb3React } from '@web3-react/core';
-
-// Components
-import { BidTable, ConnectWalletButton, OwnedDomainsTable } from 'containers';
 
 // Library
 import { ROUTES } from 'constants/routes';
 import { useDidMount } from 'lib/hooks/useDidMount';
 import { useNavbar } from 'lib/hooks/useNavbar';
+import { COPY_LABELS, TABS } from './Profile.constants';
+import { chainIdToNetworkName } from 'lib/network';
+import { useUpdateEffect } from 'lib/hooks/useUpdateEffect';
+import { useWeb3React } from '@web3-react/core';
+import { truncateWalletAddress } from 'lib/utils';
 
 // Styles
 import styles from './Profile.module.scss';
 import classNames from 'classnames/bind';
 import { ArrowLeft } from 'react-feather';
-import { truncateWalletAddress } from 'lib/utils';
 import userIcon from 'assets/user_icon.svg';
-import useNotification from 'lib/hooks/useNotification';
-import { chainIdToNetworkName } from 'lib/network';
+
 const cx = classNames.bind(styles);
 
-// Keeping these in here to reduce number of files
+/**
+ * Converts a route to a Profile route
+ * i.e., prepends the Profile route
+ * @param route to convert
+ * @returns route with profile route prepended
+ */
 const r = (route: string) => {
 	return ROUTES.PROFILE + route;
 };
-
-type Tab = {
-	title: string;
-	component: () => any;
-	location: string;
-};
-
-/**
- * NOTE: Default route will always be first tab
- */
-const TABS: Tab[] = [
-	{
-		title: 'Owned Domains',
-		component: () => <OwnedDomainsTable />,
-		location: ROUTES.OWNED_DOMAINS,
-	},
-	{
-		title: 'Your Bids',
-		component: () => <BidTable />,
-		location: ROUTES.YOUR_BIDS,
-	},
-];
 
 /**
  * User profile page for the currenctly connected user.
@@ -62,15 +49,42 @@ const Profile = () => {
 	const { account, chainId } = useWeb3React();
 
 	// React-router stuff
-	const { addNotification } = useNotification();
 	const { length: canGoBack, goBack, push, location } = useHistory();
 	const { path } = useRouteMatch();
+
 	const { setNavbarTitle } = useNavbar();
+
+	/**
+	 * Need to handle account hovers here, as we have a label
+	 * change depending on "is hovered" and "has been clicked"
+	 */
+	const [copyLabel, setCopyLabel] = useState<string>(COPY_LABELS.DEFAULT);
+	const [isAccountHovered, setIsAccountHovered] = useState<boolean>(false);
+
+	/////////////
+	// Effects //
+	/////////////
+
+	/**
+	 * Reset the copy label every time we hover/unhover
+	 */
+	useUpdateEffect(() => {
+		if (isAccountHovered) {
+			setCopyLabel(COPY_LABELS.DEFAULT);
+		}
+	}, [isAccountHovered]);
 
 	useDidMount(() => {
 		setNavbarTitle('Your Profile');
 	});
 
+	///////////////
+	// Functions //
+	///////////////
+
+	/**
+	 * Returns from profile back to where the user arrived from
+	 */
 	const onBack = () => {
 		if (canGoBack) {
 			goBack();
@@ -81,13 +95,18 @@ const Profile = () => {
 
 	/**
 	 * Copies the profile wallet address
+	 * Changes copy label to "copied"
 	 */
 	const copyAddress = () => {
 		if (account) {
 			navigator.clipboard.writeText(account);
-			addNotification('Copied address to clipboard.');
+			setCopyLabel(COPY_LABELS.COPIED);
 		}
 	};
+
+	////////////
+	// Render //
+	////////////
 
 	return (
 		<main className={styles.Container}>
@@ -96,25 +115,28 @@ const Profile = () => {
 					<ArrowLeft color="white" /> <span>My Profile</span>
 				</button>
 				{account && chainId && (
-					<button
-						className={classNames(
-							styles.Account,
-							'border-rounded',
-							'no-select',
-						)}
-						onClick={copyAddress}
-					>
-						<img alt="user icon" src={userIcon} />
-						<div className={styles.Details}>
-							<span>{chainIdToNetworkName(chainId)}</span>
-							<span>{truncateWalletAddress(account, 4)}</span>
-						</div>
-					</button>
+					<Tooltip placement="bottom-center" text={copyLabel}>
+						<button
+							className={classNames(
+								styles.Account,
+								'border-rounded',
+								'no-select',
+							)}
+							onClick={copyAddress}
+							onMouseEnter={() => setIsAccountHovered(true)}
+							onMouseLeave={() => setIsAccountHovered(false)}
+						>
+							<img alt="user icon" src={userIcon} />
+							<div className={styles.Details}>
+								<span>{chainIdToNetworkName(chainId)}</span>
+								<span>{truncateWalletAddress(account, 4)}</span>
+							</div>
+						</button>
+					</Tooltip>
 				)}
 			</div>
 			{account ? (
 				<>
-					{/* <CopyInput className={styles.Address} value={account} /> */}
 					<nav className={styles.Nav}>
 						{TABS.map((route) => (
 							<Link
